Update dao articles
