+++ src/page-components/Governance/Governance.tsx
@@ -2,14 +2,23 @@ import { GovernanceArticles } from '../../components/GovernanceArticles/Governan
 import { Hero } from './Hero/Hero';
 import { Illustration } from './Illustration/Illustration';
 import { Tokenomics } from './Tokenomics/Tokenomics';
+import { usePrismicData } from '../../lib/Prismic/components/PrismicDataProvider';
 import React from 'react';
 
 export const Governance = () => {
+    const { data } = usePrismicData();
+    const { daoArticles } = data || {};
+    const articles = daoArticles?.data?.articles;
+
     return (
         <>
             <Hero />
             <Tokenomics />
-            <GovernanceArticles sBackground="backgroundSecondaryDisabled" sPadding={{ md: '4 0', xs: '2 0' }} />
+            <GovernanceArticles
+                articles={articles as any}
+                sBackground="backgroundSecondaryDisabled"
+                sPadding={{ md: '4 0', xs: '2 0' }}
+            />
             <Illustration />
         </>
     );
